Move snapshot code to Vm in service model.
The snapshot code in automate was moved from VmOrTemplate to VMware specific class by PR https://github.com/ManageIQ/manageiq/pull/3707 when VMware was the only class that supports snapshot.
But snapshot support has been added to RHEVM and Openstack as of version Euwe.
The snapshot code in automate should be available at Vm level.

https://bugzilla.redhat.com/show_bug.cgi?id=1395175
